requirements.txt
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-api-python-client==2.45.0
twilio==7.12.0
python-dotenv==0.19.2
